server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.32
alembic==1.13.2
psycopg2-binary==2.9.9
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
python-multipart==0.0.9
python-slugify==8.0.4
APScheduler==3.10.4

# testing
pytest==8.3.3
pytest-asyncio==0.23.8
httpx==0.27.0
faker==26.0.0
bcrypt==3.2.2
